Move apps to hardcoded v1 path/directory
