+++ internal/gateway/application/service/service.go
@@ -24,8 +24,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/klog/v2"
 
-	"github.com/slackhq/spark-gateway/internal/gateway/application/handler"
 	clusterPkg "github.com/slackhq/spark-gateway/internal/gateway/cluster"
+	"github.com/slackhq/spark-gateway/internal/gateway/v1/application/handler"
 	"github.com/slackhq/spark-gateway/pkg/config"
 	"github.com/slackhq/spark-gateway/pkg/util"
