+++ internal/gateway/v1/application/handler/mockgatewayapplicationservice.go
